+++ activerecord/lib/active_record/enum.rb
@@ -157,7 +157,7 @@ module ActiveRecord
 
         # def self.statuses() statuses end
         detect_enum_conflict!(name, name.pluralize, true)
-        singleton_class.send(:define_method, name.pluralize) { enum_values }
+        singleton_class.send(:define_method, name.pluralize) { enum_values.dup }
         defined_enums[name] = enum_values
 
         detect_enum_conflict!(name, name)